+++ src/index.ts
@@ -5,6 +5,7 @@ export * from './health/index.js';
 export * from './helmet/index.js';
 export * from './properties-volume/index.js';
 export * from './csrf-token/index.js';
+export * from './redis/index.js';
 export * from './routes/index.js';
 export * from './services/index.js';
 export * from './user-state/index.js';