+++ src/interfaces/user-state-config.ts
@@ -1,0 +1,8 @@
+class UserStateConfiguration {
+  cacheKeyPrefix!: string;
+  redisClientKey!: string;
+  routePath!: string;
+  tokenClaim!: string;
+}
+
+export default UserStateConfiguration;